misctime: Use std::chrono::steady_clock instead of gettimeofday()
非推奨とされている`gettimeofday()`[1]のかわりに`std::chrono`を使います。
経過時間を測定するため標準で提供されているクロックのうちmonotonicな
タイプを選択します。
includeディレクティブの変更でコンパイルエラーが発生したためロジックを
変更していないソースファイルのヘッダーを整理しています。

[1]: https://pubs.opengroup.org/onlinepubs/9699919799/functions/gettimeofday.html

+++ src/jdlib/misctime.h
@@ -6,7 +6,6 @@
 #define _MISCTIME_H
 
 #include <string>
-#include <sys/time.h>
 #include <ctime>
 
 namespace MISC
@@ -37,8 +36,8 @@ namespace MISC
     std::string timettostr( const time_t time_from, const int mode );
 
     // 実行時間測定用
-    void start_measurement( const int id );
-    int measurement( const int id );
+    void start_measurement( const unsigned int id );
+    long long measurement( const unsigned int id );
 }
 
 #endif
